tests/factories.py
import datetime

import factory

from ads.models import Ad
from categories.models import Category
from users.models import User


class UserFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = User

    username = factory.Faker('name')
    birth_date = datetime.date(year=2010, month=10, day=12)
    email = factory.Faker('email')


class CategoryFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = Category

    name = factory.Faker('name')
    slug = factory.Faker('ean', length=8)


class AdFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = Ad

    name = factory.Faker('name')
    price = 555
    category = factory.SubFactory(CategoryFactory)
    author = factory.SubFactory(UserFactory)