pyproject.toml
[tool.poetry]
name = "homework-27"
version = "0.1.0"
description = ""
authors = ["porublevnik <porublevnik@gmail.com>"]
readme = "README.md"
packages = [{include = "homework_27"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
psycopg2-binary = "^2.9.6"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-pytest = "^0.2.0"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
